Update preferences_aboutus.xml
Switch the credits to the kyotokek tab preference. When it was basepreferencescreen the icon used was for new messages and was a patch nines. Not sure how that would of turned out but seemed like a bad idea in general. Also i properly aligned the xml so there aren't massive spaces